chore: refactor APIs
