+++ packages/storage/src/providers/s3/utils/resolveS3ConfigAndInput.ts
@@ -42,6 +42,7 @@ export const resolveS3ConfigAndInput = async ({
 		libraryOptions,
 		identityIdProvider,
 	} = config;
+	console.log(serviceOptions);
 	const { bucket, region, dangerouslyConnectToHttpEndpointForTesting } =
 		serviceOptions ?? {};
 	assertValidationError(!!bucket, StorageValidationErrorCode.NoBucket);
