+++ packages/storage/src/types/inputs.ts
@@ -85,12 +85,14 @@ export type StorageUploadDataInputWithPath<Options> =
 export interface StorageCopyInputWithKey<
 	SourceOptions extends StorageOptions,
 	DestinationOptions extends StorageOptions,
-> {
+	Options,
+> extends StorageOperationOptionsInput<Options> {
 	source: SourceOptions;
 	destination: DestinationOptions;
 }
 
-export interface StorageCopyInputWithPath {
+export interface StorageCopyInputWithPath<Options>
+	extends StorageOperationOptionsInput<Options> {
 	source: StorageOperationInputWithPath;
 	destination: StorageOperationInputWithPath;
 }
